pyproject.toml
[tool.poetry]
name = "payment-system"
version = "0.1.0"
description = "A system for payment processing and invoice generation using Stripe, uAgents, and CrewAI."

[tool.poetry.dependencies]
python = "^3.10"
stripe-agent-toolkit = "^1.0.0"
crewai = "^0.2.1"
uagents = "^0.3.0"
python-dotenv = "^0.21.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
